add readme

+++ models/product.go
@@ -4,7 +4,7 @@ type (
 	Product struct {
 		ID       string `json:"id,omitempty"`
 		Name     string `json:"name"`
-		Quantity int    `json:"quantity"`
-		Price    int    `json:"price"`
+		Quantity int64  `json:"quantity"`
+		Price    int64  `json:"price"`
 	}
 )
